Little changes

+++ src/main/java/ChartsGenerator.py
@@ -6,33 +6,37 @@ import shutil as shutil
 
 
 project_list = ["BOOKKEEPER", "OPENJPA"]
-dataset_path_list = []
+valid_project_list = []
 
 evaluation_path = "/home/lux/Documents/GitHub/ISW2_CodeMetrics/output/{name_1}/Evaluation/{name_2}_Evaluation.csv"
 box_output_image_path = "/home/lux/Documents/GitHub/ISW2_CodeMetrics/output/{name}/Charts/Box/{imageTitle}.png"
 line_output_image_path = "/home/lux/Documents/GitHub/ISW2_CodeMetrics/output/{name}/Charts/Line/{imageTitle}.png"
-image_directory = "/home/lux/Documents/GitHub/ISW2_CodeMetrics/output/{name}/Charts/"
+comparison_image_path = "/home/lux/Documents/GitHub/ISW2_CodeMetrics/output/{name}/Charts/"
+image_directory = "/home/lux/Documents/GitHub/ISW2_CodeMetrics//output/{name}/Charts/"
 
 def main() :
     init_directories()
-    analyze_all_projects()
+    analyze_all_projetcs()
 
 
 def init_directories() :
     for project_name in project_list :
-        dataset_path_list.append(evaluation_path.format(name_1 = project_name, name_2 = project_name))
+        path = evaluation_path.format(name_1 = project_name, name_2 = project_name)
+        if (os.path.exists(path)) :
+            valid_project_list.append(project_name)
+            directory = image_directory.format(name = project_name)
+            if (os.path.isdir(directory)) :
+                shutil.rmtree(path = directory)
+            os.mkdir(directory)
+            os.mkdir(directory + "/" + "Box")
+            os.mkdir(directory + "/" + "Line")
+            os.mkdir(directory + "/" + "Comparison")
 
-    for project_name in project_list :
-        directory = image_directory.format(name = project_name)
-        if (os.path.isdir(directory)) :
-            shutil.rmtree(path = directory)
-        os.mkdir(directory)
-        os.mkdir(directory + "/" + "Box")
-        os.mkdir(directory + "/" + "Line")
 
-def analyze_all_projects() :
-    for project_name in project_list :
-        analyze_project(project_name)
+
+def analyze_all_projetcs() :
+    for projectName in valid_project_list :
+        analyze_project(projectName)
 
 
 def analyze_project(project_name) :
@@ -42,8 +46,8 @@ def analyze_project(project_name) :
 
     versions = dataset["TrainingRelease"].drop_duplicates().values
     classifiers = dataset["ClassifierName"].drop_duplicates().values
-    filters = dataset["FilterName"].drop_duplicates().values
-    samplers = dataset["SamplerName"].drop_duplicates().values
+    filters = dataset["Filter"].drop_duplicates().values
+    samplers = dataset["Sampler"].drop_duplicates().values
     sensitive = dataset["SensitiveLearning"].drop_duplicates().values
 
     for filter in filters :
@@ -51,22 +55,73 @@ def analyze_project(project_name) :
             for is_sensitive in sensitive :
                 if (sampler != "NotSet" and is_sensitive) :
                     continue
+
                 box_plot_data(project_name, dataset, classifiers, filter, sampler, is_sensitive)
                 line_plot_data(project_name, dataset, versions, classifiers, filter, sampler, is_sensitive)
+
+    metric_list = ["Precision", "Recall", "ROC_AUC"]
+    for metric in metric_list :
+        comparison_box_plot(project_name, dataset, classifiers, filters, samplers, sensitive, metric)
+
+
+def comparison_box_plot(project_name, dataset, classifier_list, filter_list, sampler_list, sensitive_list, metric) :
+    nCols = 0
+    for filter in filter_list :
+        for sampler in sampler_list :
+            for is_sensitive in sensitive_list :
+
+                if (is_sensitive and sampler != "NotSet") :
+                    continue
+
+                nCols = nCols + 1
+
+    figure, axis = plt.subplots(nrows = 1, ncols = nCols, sharey = "row")
+    figure.set_size_inches(20,10)
+    figure.set_tight_layout(tight = {"w_pad" : -0.75})
+    #figure.subplots_adjust(wspace = 0)
+
+    index = 0
+    for filter in filter_list :
+        for sampler in sampler_list :
+            for is_sensitive in sensitive_list :
+
+                if (is_sensitive and sampler != "NotSet") :
+                    continue
+
+                precision_data_list = []
+
+                for classifier in classifier_list :
+                    precision_data = get_data(dataset, None, classifier, filter, sampler, is_sensitive)[metric]
+                    precision_data = precision_data[precision_data.notnull()]
+                    precision_data_list.append(precision_data)
+
+                axis[index].boxplot(precision_data_list)
+                axis[index].set_xticklabels(classifier_list, rotation = 60)
+                axis[index].set_ylim(-0, 1)
+                axis[index].yaxis.grid(linestyle = '--', linewidth = 0.75)
+                axis[index].set_yticks(np.arange(-0.05,1.1, 0.05))
+                axis[index].set_title("Filter = " + str(filter) + "\nSampling = " + sampler + "\nSensitive = " + str(is_sensitive))
+
+                index += 1
+
+    output_path = comparison_image_path.format(name = project_name, imageTitle =project_name + "_" + metric + "_Comparison")
+
+    figure.suptitle(project_name + " " + metric + " Comparison")
+    figure.savefig(output_path)
 
 
 
 def line_plot_data(project_name, dataset, version_list, classifier_list, filter, sampler, is_sensitive) :
 
     figure, axis = plt.subplots(nrows = 1, ncols = 3)
-    figure.set_size_inches(16,5)
-    figure.subplots_adjust(hspace=0.3)
+    figure.set_size_inches(15,5)
+    figure.set_tight_layout(tight = {"h_pad" : 0.3})
 
-    title_string = "LinePlot.Filter:{filterName}-Sampler:{samplerName}-IsSensitive:{sensitive}"
-    title_string = title_string.format(filter_name = filter, sampler_name = sampler, sensitive = is_sensitive)
+    title_string = "LinePlot-(Filter = {filterName})-(Sampler = {samplerName})-(IsSensitive = {sensitive})"
+    title_string = title_string.format(filterName = filter, samplerName = sampler, sensitive = is_sensitive)
     figure.suptitle(title_string)
 
-    image_path = line_output_image_path.format(name = project_name, image_title = title_string)
+    image_path = line_output_image_path.format(name = project_name, imageTitle = title_string)
 
     for index in range(0, len(classifier_list)) :
 
@@ -88,26 +143,27 @@ def line_plot_data(project_name, dataset, version_list, classifier_list, filter,
         axis[2].set_title("ROC_AUC")
 
     for i in range(0, 3) :
-        axis[i].legend()
+        axis[i].legend(loc = 'upper left')
         axis[i].grid()
-        axis[i].set_yticks(np.arange(0,1.1, 0.1))
+        axis[i].set_yticks(np.arange(0, 1.4, 0.1))
         axis[i].set_xticks(np.arange(0, len(version_list), 1))
 
 
-    plt.savefig(image_path)
+    # figure.legend(labels = classifierList)
+    figure.savefig(image_path)
     #plt.show()
 
 
 
 def box_plot_data(project_name, dataset, classifier_list, filter, sampler, cost_sensitive) :
     figure, axis = plt.subplots(2, 2)
-    figure.set_size_inches(16,9)
-    figure.subplots_adjust(hspace=0.3)
+    figure.set_size_inches(9,9)
+    #figure.subplots_adjust(hspace=0.3)
+    figure.set_tight_layout(tight = {"h_pad" : 0.3})
 
-    title_string = "BoxPlot.Filter:{filterName}-Sampler:{samplerName}-IsSensitive:{sensitive}"
+    title_string = "BoxPlot-(Filter = {filterName})-(Sampler = {samplerName})-(IsSensitive = {sensitive})"
     title_string = title_string.format(filterName = filter, samplerName = sampler, sensitive = cost_sensitive)
     figure.suptitle(title_string)
-
 
     image_path = box_output_image_path.format(name = project_name, imageTitle = title_string)
 
@@ -138,10 +194,15 @@ def box_plot_data(project_name, dataset, classifier_list, filter, sampler, cost_
 
     for i in range(0,2) :
         for j in range(0,2) :
-            axis[i,j].set_xticklabels(classifier_list)
+            axis[i,j].set_xticklabels(classifier_list, rotation = 15)
             axis[i,j].set_ylim(-0.1, 1)
             axis[i,j].yaxis.grid()
-            axis[i,j].set_yticks(np.arange(0,1.1, 0.1))
+            axis[i,j].set_yticks(np.arange(-0.1,1.1, 0.1))
+
+
+            if (i == 1 and j == 1) :
+                axis[i,j].set_ylim(-0.5, 1)
+                axis[i,j].set_yticks(np.arange(-1, 1.1, 0.2))
 
 
     axis[0,0].boxplot(recall_list)
@@ -156,7 +217,7 @@ def box_plot_data(project_name, dataset, classifier_list, filter, sampler, cost_
     axis[1,1].boxplot(kappa_list)
     axis[1,1].set_title("Kappa")
 
-    plt.savefig(image_path)
+    figure.savefig(image_path)
     #plt.show()
 
 
@@ -167,9 +228,9 @@ def get_data(dataset, version, classifier, my_filter, sampler, is_sensitive) :
     if (classifier != None) :
         filtered_dataset = filtered_dataset[(filtered_dataset["ClassifierName"] == classifier)]
     if (my_filter != None) :
-        filtered_dataset = filtered_dataset[(filtered_dataset["FilterName"] == my_filter)]
+        filtered_dataset = filtered_dataset[(filtered_dataset["Filter"] == my_filter)]
     if (sampler != None) :
-        filtered_dataset = filtered_dataset[(filtered_dataset["SamplerName"] == sampler)]
+        filtered_dataset = filtered_dataset[(filtered_dataset["Sampler"] == sampler)]
     if (is_sensitive != None) :
         filtered_dataset = filtered_dataset[filtered_dataset["SensitiveLearning"] == is_sensitive]
